Move block validation code to BlockChain
